+++ deploy/k8s/deploy-accelerator.sh
@@ -20,9 +20,12 @@ if [ ! -z "$uninstall" ]; then
        exit 1
    fi
    echo "Uninstalling MEC accelerator"
-   if kubectl get BrokerListener -n azure-iot-operations | grep -q 'mec-listener'; then
-      kubectl delete -f ./E4K/
-   fi
+   if [ kubectl get ns | grep -q 'azure-iot-operations' ]; then
+        if kubectl get BrokerListener -n azure-iot-operations | grep -q 'mec-listener'; then
+            kubectl delete -f ./E4K/
+        fi
+    fi
+  
    if kubectl get pods -n mec-accelerator | grep -q 'mosquitto'; then
       kubectl delete -f ./mosquitto/
    fi
